Replace make.bat with F# script
We were using an F# script before for building (make) but it got
replaced with autotools because we couldn't rely on Ubuntu's buggy
fsharpi. We can, however, assume that FSI works in Windows, so we
resurrect it for this OS only.

This way I avoid the nightmares I'm having about batch files syntax
that I was trying to work with in the make.bat file.